Refactor the actions/utils package and extract action tag to its own submodule
